requirements.txt
gradio
pytest
coverage
python-dotenv
requests